String manipulation for file names in separate translation unit
