+++ src/conn.c
@@ -263,18 +263,3 @@ curl_off_t getSize(char* filename, char* url)
     
 }
 
-// Getting filename from URL
-char* filenameFromURL(char* URLString)
-{
-    unsigned int offset = 0;
-    
-    for(unsigned int i = 0; i < strlen(URLString); i++)
-    {
-        if(URLString[i] == '/')
-        {
-            offset = i + 1;
-        }
-    }
-
-    return URLString + offset;
-}
