+++ src/filestring.h
@@ -1,0 +1,9 @@
+#ifndef FILESTRING_H
+#define FILESTRING_H
+
+
+#include <string.h>
+
+char* filenameFromURL(char* URLString);
+
+#endif
